refactor: Extract test utilities to `util.go`
